+++ cli/auth/login/register.ts
@@ -3,6 +3,7 @@ import { setSessionToken, getSessionToken } from "lib/cli-config"
 import delay from "delay"
 import { getRegistryApiKy } from "lib/registry-api/get-ky"
 import type { EndpointResponse } from "lib/registry-api/endpoint-types"
+import { setupNpmrc } from "../setup-npmrc/setup-npmrc"
 
 export const registerAuthLogin = (program: Command) => {
   // Define the login action once to share between both commands
@@ -73,6 +74,9 @@ export const registerAuthLogin = (program: Command) => {
 
     setSessionToken(session.token)
     console.log("\nReady to use!")
+
+    // Setup npmrc with the new session token in global level ~/.npmrc file
+    setupNpmrc(session.token)
   }
 
   // Register the auth login subcommand
