+++ cli/main.ts
@@ -24,7 +24,7 @@ import { registerRemove } from "./remove/register"
 import { registerBuild } from "./build/register"
 import { registerSnapshot } from "./snapshot/register"
 import { registerSetup } from "./setup/register"
-import { registerSetupNpmrc } from "./setup-npmrc/register"
+import { registerAuthSetupNpmrc } from "./auth/setup-npmrc/register"
 import { registerConvert } from "./convert/register"
 import { registerSimulate } from "./simulate/register"
 import { registerInstall } from "./install/register"
@@ -46,6 +46,7 @@ registerAuthLogout(program)
 registerAuthPrintToken(program)
 registerAuthSetToken(program)
 registerAuthWhoami(program)
+registerAuthSetupNpmrc(program)
 
 registerConfig(program)
 registerConfigPrint(program)
@@ -58,7 +59,6 @@ registerAdd(program)
 registerRemove(program)
 registerSnapshot(program)
 registerSetup(program)
-registerSetupNpmrc(program)
 registerInstall(program)
 registerUpgradeCommand(program)
